Updates to TouchOSC and Launch_Control scripts for 16 knob support
LC now has a default mode where the 8 buttons are track select + view selected track, with navigation buttons for magic box (box does not appear).

TouchOSC is the same as previous iteration for now, with the same pad-based navigational components to support the use of the same script with my FaderFox EC4.